replace manual link status with NavLink
