+++ src/ui/index.css
@@ -134,7 +134,7 @@ tr.highlight td {
 
 .menu-item {
   padding: 10px;
-  border-radius: 8px 8px 0 0;
+  border-radius: 8px;
 }
 
 .menu-item.on {
